refactor: refactor: enhance modular component structure and layout architecture
