+++ src/Admin/components/AdminPanelWrapper/styles.module.scss
@@ -1,0 +1,55 @@
+$breakpoint-desktop: 900px;
+
+.container {
+    display: flex;
+    flex-direction: column;
+    gap: var(--givewp-spacing-4);
+    min-height: 100%;
+    flex: 1;
+}
+
+.aboveSection {
+    width: 100%;
+    flex-shrink: 0;
+    display: flex;
+    gap: var(--givewp-spacing-4);
+    
+    // Make child elements (StatWidgets) grow equally
+    > * {
+        flex: 1;
+    }
+    
+    @media (max-width: $breakpoint-desktop) {
+        flex-direction: column;
+        gap: var(--givewp-spacing-3);
+        
+        > * {
+            flex: none;
+        }
+    }
+}
+
+.grid {
+    display: grid;
+    grid-template-columns: 2fr 1fr;
+    gap: var(--givewp-spacing-4);
+    flex: 1;
+    align-items: start;
+
+    @media (max-width: $breakpoint-desktop) {
+        grid-template-columns: 1fr;
+        gap: var(--givewp-spacing-6);
+    }
+}
+
+.leftColumn, .rightColumn {
+    display: flex;
+    flex-direction: column;
+    gap: var(--givewp-spacing-6);
+    height: fit-content;
+}
+
+.belowSection {
+    width: 100%;
+    flex-shrink: 0;
+}
