+++ src/Donors/resources/admin/components/DonorDetailsPage/Tabs/Overview/DonorContributions/styles.module.scss
@@ -1,0 +1,11 @@
+.contributionsCard {
+    display: flex;
+    flex-direction: column;
+
+    > div:last-child {
+        flex: 1;
+        width: 100%;
+        height: 100%;
+        min-height: 300px;
+    }
+}
