+++ src/Donors/resources/admin/components/DonorDetailsPage/Tabs/Overview/DonorSummary/styles.module.scss
@@ -1,0 +1,85 @@
+$breakpoint-desktop: 900px;
+
+.card {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: flex-start;
+    padding: 1rem 1.5rem 1.5rem;
+    border-radius: 8px;
+    border: solid 1px #f3f4f6;
+    background-color: var(--givewp-shades-white);
+    width: 100%;
+    box-sizing: border-box;
+}
+
+.summaryCard {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+    border-bottom: solid 1px #f3f4f6;
+    background-color: var(--givewp-shades-white);
+
+    @media (max-width: $breakpoint-desktop) {
+        flex-wrap: wrap;
+    }
+
+    .summaryCardLabel {
+        flex-grow: 1;
+        font-size: 1rem;
+        font-weight: 500;
+        line-height: 1.5;
+        color: #4b5563;
+        padding: var(--givewp-spacing-5) 0 var(--givewp-spacing-3) 0;
+        margin: 0;
+    }
+
+    .summaryCardValue {
+        font-size: 1rem;
+        font-weight: 600;
+        line-height: 1.5;
+        color: #1f2937;
+        padding: var(--givewp-spacing-5) 0 var(--givewp-spacing-3) 0;
+
+        &.pill {
+            background-color: #f9fafb;
+            padding: 0.25rem 0.75rem;
+            border-radius: 9999px;
+            color: #060c1a;
+            border: solid 1px #374151;
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            margin: var(--givewp-spacing-5) 0 var(--givewp-spacing-3) 0;
+            font-size: 0.75rem;
+        }
+    }
+
+    .summaryCardValues {
+        display: flex;
+        flex-direction: column;
+        gap: var(--givewp-spacing-1);
+        margin: 0;
+        padding: var(--givewp-spacing-5) 0 var(--givewp-spacing-3) 0;
+
+        p {
+            margin: 0;
+
+            &:first-child {
+                font-size: 1rem;
+                font-weight: 600;
+                line-height: 1.5;
+                color: #459948;
+                text-align: right;
+            }
+
+            &:last-child {
+                font-size: 0.875rem;
+                font-weight: 500;
+                line-height: 1.43;
+                color: #4b5563;
+            }
+        }
+    }
+}
