+++ src/Views/Components/AdminDetailsPage/index.tsx
@@ -30,6 +30,7 @@ import TabPanels from './Tabs/TabPanels';
 import DefaultPrimaryActionButton from './DefaultPrimaryActionButton';
 import AdminSection, { AdminSectionField, AdminSectionsWrapper } from './AdminSection';
 
+
 import './store';
 
 /**
